DataCollector/requirements.txt
mysql-connector-python
yfinance
confluent-kafka==1.9.2
prometheus_client


